+++ src/main/java/com/demosocket/manager/repository/SystemRepository.java
@@ -3,5 +3,5 @@ package com.demosocket.manager.repository;
 import com.demosocket.manager.model.System;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SystemRepository extends JpaRepository<System, Integer> {
+public interface SystemRepository extends JpaRepository<System, Long> {
 }
